RE-3159 - Fix: Zod, dispatcher toutes les erreurs en même temps
